+++ docs/changelog.md
@@ -1,8 +1,9 @@
 Changelog
 =========
 
-### 2023-03-17: v.0.2.0
-- initial release
+### 2023-11-23: v.0.2.0
+- fixed lichess "opponent gone" behaviour, mate and stalemate handling
+- upgraded to latest PyTorch Lightning version 2.1.2
 
 ## 2023-03-17: v.0.1.0
 - initial release